+++ client/changes.md
@@ -120,7 +120,7 @@ if (createdBy === me || isEditable) allowEdit
 - move all business logic to service XX
 - ~~use tailwind class instead of [hidden] --~~
 
-## 25-2-2024
+### 25-2-2024
 
 - fix expression has changed error in users-component (console error)
 - fix undefined error in users-component (console error)
@@ -128,4 +128,6 @@ if (createdBy === me || isEditable) allowEdit
 - fix deleting files
 - fix shared by (show name, not id...)
 - move loading to a separate interceptor, and create only 1 method to toggle loading
+- move all business logic to service
+- remove createdBy from client
 - move all business logic to service